+++ apps/ced-browser-be/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ced-browser-be
 
+## 0.1.5
+
+### Patch Changes
+
+- 6f5c703: add profile info and make user info optional
+
 ## 0.1.4
 
 ### Patch Changes